+++ CHANGELOG.md
@@ -89,12 +89,13 @@
 ### Removed
 
 * nostr: remove `NegentropyErrorCode` ([Yuki Kishimoto])
-* sdk: remove `nwc` dep ([Yuki Kishimoto])
 * database: remove `lru`, `thiserror` and `tracing` deps ([Yuki Kishimoto])
 * lmdb: remove `thiserror` and `tracing` deps ([Yuki Kishimoto])
 * indexeddb: remove `thiserror` and `tracing` deps ([Yuki Kishimoto])
 * zapper: remove `thiserror` dep ([Yuki Kishimoto])
 * pool: remove `thiserror` and `tokio-stream` deps ([Yuki Kishimoto])
+* pool: remove minimum interval constraint in `RelayOptions::retry_interval`
+* sdk: remove `nwc` dep ([Yuki Kishimoto])
 * nwc: remove `thiserror` dep and unnecessary `Error::Zapper` variant ([Yuki Kishimoto])
 * ffi: remove `MockRelay` ([Yuki Kishimoto])
 * ffi: remove `RawEvent` and `RawEventRecord` ([Yuki Kishimoto])